tests/wpt/tests/css/css-conditional/container-queries/fragmented-container-001.html

<!doctype html>
<title>CSS Container Queries Test: Query fragmented inline-size container</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#size-container">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #multicol {
    width: 400px;
    column-count: 2;
    column-fill: auto;
    column-gap: 0;
    height: 100px;
  }
  #float {
    float: left;
    width: 100px;
    height: 50px;
  }
  #container {
    container-type: inline-size;
    display: flow-root;
    height: 200px;
  }
  #first-child {
    break-after: column;
  }
  @container (width = 100px) {
    #first-child { color: green; }
    #second-child { color: green; }
  }
</style>
<div id="multicol">
  <div id="float"></div>
  <div id="container">
    <div id="first-child"></div>
    <div id="second-child"></div>
  </div>
</div>
<script>
  setup(() => assert_implements_container_queries());

  test(() => {
    const green = "rgb(0, 128, 0)";
    assert_equals(getComputedStyle(document.querySelector("#first-child")).color, green);
    assert_equals(getComputedStyle(document.querySelector("#second-child")).color, green);
  }, "Children of fragmented inline-size container should match inline-size of first fragment");
</script>
